requirements.txt
streamlit
opencv-python
numpy
Pillow
pytesseract
pandas
matplotlib
prettytable
pymongo
math
io
xlsxwriter